ASoC: ux500: Remove redundant casts
JIRA: https://issues.redhat.com/browse/RHEL-80681

commit fa24fdc8ae9e6f6fe7a4f7676a4d8c14433a86c0
Author: Tang Bin <tangbin@cmss.chinamobile.com>
Date: Thu Nov 7 15:54:40 2024 +0800

    ASoC: ux500: Remove redundant casts

    In the function ux500_msp_drv_probe, the 'int' type cast
    in front of the PTR_ERR() macro is redundant, thus remove
    it.

    Link: https://patch.msgid.link/20241107075440.2770-1-tangbin@cmss.chinamobile.com

+++ sound/soc/ux500/ux500_msp_dai.c
@@ -761,7 +761,7 @@ static int ux500_msp_drv_probe(struct platform_device *pdev)
 
 	drvdata->reg_vape = devm_regulator_get(&pdev->dev, "v-ape");
 	if (IS_ERR(drvdata->reg_vape)) {
-		ret = (int)PTR_ERR(drvdata->reg_vape);
+		ret = PTR_ERR(drvdata->reg_vape);
 		dev_err(&pdev->dev,
 			"%s: ERROR: Failed to get Vape supply (%d)!\n",
 			__func__, ret);
@@ -771,7 +771,7 @@ static int ux500_msp_drv_probe(struct platform_device *pdev)
 
 	drvdata->pclk = devm_clk_get(&pdev->dev, "apb_pclk");
 	if (IS_ERR(drvdata->pclk)) {
-		ret = (int)PTR_ERR(drvdata->pclk);
+		ret = PTR_ERR(drvdata->pclk);
 		dev_err(&pdev->dev,
 			"%s: ERROR: devm_clk_get of pclk failed (%d)!\n",
 			__func__, ret);
@@ -780,7 +780,7 @@ static int ux500_msp_drv_probe(struct platform_device *pdev)
 
 	drvdata->clk = devm_clk_get(&pdev->dev, NULL);
 	if (IS_ERR(drvdata->clk)) {
-		ret = (int)PTR_ERR(drvdata->clk);
+		ret = PTR_ERR(drvdata->clk);
 		dev_err(&pdev->dev,
 			"%s: ERROR: devm_clk_get failed (%d)!\n",
 			__func__, ret);
